Create directory tree structure Ansible
The script searches the initial / etc / ansible directory to see if the roles directory exists. If not, he created it. Ask for which role you want to create this tree. Then created the tree structure necessary for this role. To finish he created the default yaml files in the different directories