Update auto_LAMP_stack.sh
This script allows you to directly install a functional APACHE-based web solution. Interactive the script asks the user to enter information necessary for the creation of the database. The only thing left at the end of the script is to create a configuration for the VirtualHost or modify the default configuration file.